requirements.txt
streamlit
openai
groq
newsapi-python
python-dotenv